Create lib/logger.
So we can write it's logic once and import it where we need it.
