+++ lib/jsonld_serializers.js
@@ -5,7 +5,7 @@ var lexicon = require('nypl-registry-utils-lexicon')
 
 var util = require('./util.js')
 const config = require('config')
-
+const logger = require('./logger')
 const PACK_DELIM = '||'
 
 // Base class for all serialized types
@@ -218,7 +218,7 @@ class ResourceSerializer extends JsonLdItemSerializer {
   }
 
   static serialize (resp, options) {
-    console.log('ResourceSerializer#serialize', resp)
+    logger.debug('ResourceSerializer#serialize', resp)
     return (new ResourceSerializer(resp, options)).format()
   }
 }
